mcp-server/requirements.txt
mcp[cli]>=1.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.2.1
requests>=2.31.0